+++ lib/dicey.rb
@@ -5,6 +5,7 @@ module Dicey
   # General error for Dicey.
   class DiceyError < StandardError; end
 
+  Dir["dicey/mixins/*.rb", base: __dir__].each { require_relative _1 }
   Dir["dicey/*.rb", base: __dir__].each { require_relative _1 }
   Dir["dicey/output_formatters/*.rb", base: __dir__].each { require_relative _1 }
   Dir["dicey/sum_frequency_calculators/*.rb", base: __dir__].each { require_relative _1 }